LayoutTests/platform/mac-wk1/fast/loader/window-open-to-invalid-url-disallowed-expected.txt

main frame - didFinishDocumentLoadForFrame
main frame - willPerformClientRedirectToURL: http://A=a%25B=b
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
main frame - didCancelClientRedirectForFrame
Tests that we do not open a new window to an invalid URL. This test PASSED if you see an entry in the dumped frame load callbacks of the form: "willPerformClientRedirectToURL: http://A=a%B=b" followed by "didCancelClientRedirectForFrame".

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS win.testRunner.didCancelClientRedirect became true
PASS successfullyParsed is true

TEST COMPLETE
